fix: rename `AppInit` as `createApp`
